SDK for session
