+++ sdk/python/ragflow/__init__.py
@@ -4,4 +4,5 @@ __version__ = importlib.metadata.version("ragflow")
 
 from .ragflow import RAGFlow
 from .modules.dataset import DataSet
-from .modules.chat_assistant import Assistant
+from .modules.assistant import Assistant
+from .modules.session import Session
